Move classes to domain package
