+++ composeApp/src/commonMain/kotlin/com/schuetz/agents/domain/LLM.kt
@@ -1,4 +1,4 @@
-package com.schuetz.agents.agent
+package com.schuetz.agents.domain
 
 interface LLM {
     suspend fun prompt(message: String): String
